+++ src/main/java/de/johni0702/minecraft/bobby/mixin/GameOptionsMixin.java
@@ -8,6 +8,8 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
+import org.spongepowered.asm.mixin.injection.Slice;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(GameOptions.class)
@@ -20,5 +22,16 @@ public class GameOptionsMixin {
         if (Bobby.getInstance().isEnabled()) {
             ci.setReturnValue(this.viewDistance.getValue());
         }
+    }
+
+    @ModifyArg(
+            method = "<init>",
+            slice = @Slice(from = @At(value = "CONSTANT", args = "stringValue=options.renderDistance")),
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/client/option/SimpleOption$ValidatingIntSliderCallbacks;<init>(II)V", ordinal = 0),
+            index = 1
+    )
+    private int considerBobbyMaxRenderDistanceSetting(int vanillaSetting) {
+        int bobbySetting = Bobby.getInstance().getConfig().getMaxRenderDistance();
+        return Math.max(vanillaSetting, bobbySetting);
     }
 }